Performs direct login (when enabled)
When `autoLogin` is set for an application, the Worker will now catch a
GET towards the configured `loginPath` and perform a `POST` directly to
the server, handing back a session Cookie directly to the client.
Setting `autoLogin` to false will prevent this from happening and will
still rely on the JavaScript injection.